requirements.txt
SQLAlchemy>=1.4
psycopg2-binary
python-dotenv
tabulate
fastapi
uvicorn[standard]
pydantic
requests
pytest
httpx
